restructured using factory pattern
