+++ application/accounts/views.py
@@ -4,7 +4,7 @@ from flask import Blueprint, flash, g, redirect, render_template, request, sessi
 from werkzeug.security import check_password_hash
 from werkzeug.security import generate_password_hash
 from application import db
-from application.models.auth import User
+from application.models import User
 
 bp = Blueprint("auth", __name__, url_prefix="/auth")
