+++ application/models/orders.py
@@ -2,18 +2,6 @@ from sqlalchemy import Column, Integer, String, Float, ForeignKey
 from sqlalchemy.orm import relationship
 from application import db
 
-class Product(db.Model):
-    __tablename__ = 'products'
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    admin_id = Column(Integer, ForeignKey('users.id'), nullable=False)
-    name = Column(String, unique=True, nullable=False)
-    price = Column(Float, nullable=False)
-    stock = Column(Integer, nullable=False, default=0)
-    description = Column(String, nullable=False)
-    image = Column(String, nullable=False)
-
-    admin = relationship('User', backref='products')
-
 class Orders(db.Model):
     __tablename__ = 'orders'
     id = Column(Integer, primary_key=True, autoincrement=True)
@@ -42,5 +30,4 @@ class OrderLine(db.Model):
     buyer_id = Column(Integer, ForeignKey('users.id'), primary_key=True)
     qty = Column(Integer, nullable=False)
     qty_total_price = Column(Float, nullable=False)
-
     buyer = relationship('User', backref='order_lines')
