+++ application/templates/header.html
@@ -19,12 +19,12 @@
                         {% endif %}
 
                         {% if g.user %}
-                        <li><a href="{{ url_for('store.cart') }}" class="text-white">
+                        <li><a href="{{ url_for('orders.cart') }}" class="text-white">
                                 My Cart</a></li>
                         <li>
                             <hr class="dropdown-divider">
                         </li>
-                        <li><a href="{{ url_for('store.cart') }}" class="text-white">
+                        <li><a href="{{ url_for('orders.cart') }}" class="text-white">
                             My Account</a></li>
                         <li><a href="{{ url_for('auth.logout') }}" class="text-white">Log Out</a></li>
                         {% else %}
