+++ application/templates/products/cart.html
@@ -26,10 +26,10 @@
 
       <div class="col p-4 d-flex flex-column position-static">
         <strong class="d-inline-block mb-2 text-primary">Delivery 1-2 Days</strong>
-        <h3 class="mb-0"><a href="{{ url_for('store.read', id=product['id']) }}">{{ product['name'] }}</a></h3>
+        <h3 class="mb-0"><a href="{{ url_for('products.read', id=product['id']) }}">{{ product['name'] }}</a></h3>
         <p class="card-text mb-auto">€ {{ product.price }} x {{ product.qty }} = € {{ product.qty_total_price }}</p>
 
-        <a href="{{ url_for('store.cart_remove_item', index=product['index']) }}">Remove 1 Qty</a>
+        <a href="{{ url_for('orders.cart_remove_item', index=product['index']) }}">Remove 1 Qty</a>
 
       </div>
 
@@ -48,7 +48,7 @@
     <hr class="my-3" />
 		<h3 class="display-6"> Total payment: € {{ total_payment |round(2, 'floor') }}</h3>
 
-		<p> <a href="{{ url_for('store.checkout') }}" class="btn btn-lg btn-danger my-3">Proceed to checkout</a></p>
+		<p> <a href="{{ url_for('orders.checkout') }}" class="btn btn-lg btn-danger my-3">Proceed to checkout</a></p>
     {% else %}
     <h3>You have 0 items in your cart.</h3>
     {% endif %}
