+++ tests/unit/test_store.py
@@ -6,7 +6,6 @@ def test_index(test_app_with_db):
     response = test_app_with_db.get('/')
     assert response.status_code == 200
 
-
 def test_get_product_by_admin(test_app_with_db):
     pass
 
@@ -31,7 +30,6 @@ def test_get_product_incorrect_id(test_app_with_db):
     response = test_app_with_db.get('/product/999')
     assert response.status_code == 404
     assert response.json['detail'] == 'Product not found'
-
 
 def test_handle_cart(test_app_with_db):
     pass
